Bugfix when no tube connected and output improvement
When no tube was connected and no counting occurred, it would wait for a
tick to happen before acceping any button presses, which is really
awkward behavior. Fixed in current revision. Also added the feature that
the user can tell the Geiger counter via RS232 that unconditionally data
shall be output even when no counting occurs. This allows checking if
the device is still connected (but for some reason not counting).